added a searcher skeleton
