+++ shoebox/app/com/keepit/search/index/Indexer.scala
@@ -28,7 +28,7 @@ class Indexer[T](indexDirectory: Directory, indexWriterConfig: IndexWriterConfig
 
   lazy val indexWriter = new IndexWriter(indexDirectory, indexWriterConfig)
 
-  var readerContext: Option[(IndexReader, Array[Long])] = None
+  var searcher: Option[Searcher] = None
   
   def doWithIndexWriter(f: IndexWriter=>Unit) = {
     try {
@@ -60,58 +60,18 @@ class Indexer[T](indexDirectory: Directory, indexWriterConfig: IndexWriterConfig
   
   def numDocs = indexWriter.numDocs()
   
-  def initReaderContext {
-    if (IndexReader.indexExists(indexDirectory)) refreshReaderContext
+  def initSearcher {
+    if (IndexReader.indexExists(indexDirectory)) refreshSearcher
   }
   
-  def refreshReaderContext {
-    val reader = readerContext match {
-      case Some((reader, _)) => 
-        IndexReader.openIfChanged(reader) // this may return null
+  def refreshSearcher {
+    val reader = searcher match {
+      case Some(searcher) => 
+        IndexReader.openIfChanged(searcher.indexReader) // this may return null
       case None =>
         if (IndexReader.indexExists(indexDirectory)) IndexReader.open(indexDirectory)
         else null
     }
-    if (reader != null) readerContext = Some((reader, buildIdMapper(reader))) // lucene may return a null reader.
-  }
-  
-  def buildIdMapper(reader: IndexReader) = {
-    val maxDoc = reader.maxDoc();
-    val idArray = new Array[Long](maxDoc); 
-    val payloadBuffer = new Array[Byte](8)
-    val tp = reader.termPositions(Indexer.idPayloadTerm);
-    try {
-      var idx = 0;
-      while (tp.next()) {
-        val doc = tp.doc();
-        while (idx < doc) {
-          idArray(idx) = Indexer.DELETED_ID; // fill the gap
-          idx += 1
-        }
-        tp.nextPosition();
-        tp.getPayload(payloadBuffer, 0)
-        val id = bytesToLong(payloadBuffer)
-        idArray(idx) = id
-        idx += 1
-      }
-      while(idx < maxDoc) {
-        idArray(idx) = Indexer.DELETED_ID; // fill the gap
-        idx += 1
-      }
-    } finally {
-      tp.close();
-    }
-    idArray
-  }
-  
-  private def bytesToLong(bytes: Array[Byte]): Long = {
-    ((bytes(0) & 0xFF).toLong) |
-    ((bytes(1) & 0xFF).toLong <<  8) |
-    ((bytes(2) & 0xFF).toLong << 16) |
-    ((bytes(3) & 0xFF).toLong << 24) |
-    ((bytes(4) & 0xFF).toLong << 32) |
-    ((bytes(5) & 0xFF).toLong << 40) | 
-    ((bytes(6) & 0xFF).toLong << 48) |
-    ((bytes(7) & 0xFF).toLong << 56)
+    if (reader != null) searcher = Some(new Searcher(reader, ArrayIdMapper(reader))) // lucene may return a null reader.
   }
 }
